Move the global wr function
