+++ main.go
@@ -3,12 +3,7 @@ package main
 import (
 	"./Client"
 	"./Data"
-	"./EnvPath"
 	"./Log"
-	"bufio"
-	"encoding/json"
-	"io"
-	"os"
 	"runtime"
 	"time"
 )
@@ -33,7 +28,7 @@ func initAll() error {
 	if err != nil {
 		return err
 	}
-	err = readGlobal()
+	err = Data.ReadGlobal(global)
 	if err != nil {
 		return err
 	}
@@ -95,44 +90,6 @@ func main() {
 	}
 	ret = start()
 	defer exit(ret)
-}
-
-func readGlobal() (err error) {
-	var data = make(map[string]string)
-	filePath := envpath.GetAppDir()
-	filePath += "/YoC.info"
-	file, err := os.OpenFile(filePath, os.O_CREATE|os.O_RDONLY, os.ModePerm)
-	scanner := bufio.NewReader(file)
-	bytes, err := scanner.ReadBytes('\n')
-	if err != nil && err != io.EOF {
-		return err
-	}
-	err = json.Unmarshal(bytes, &data)
-	if err != nil {
-		return err
-	}
-	if ps, ok := data["Password"]; !ok || ps == "" {
-		data["Password"] = "YoCProject"
-	}
-	if data["Version"] == global["Version"] {
-		global = data
-	} else {
-		data["Version"] = global["Version"]
-		global = data
-		file, err = os.OpenFile(filePath, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, os.ModePerm)
-		if err != nil {
-			return err
-		}
-		globalInfo, err := json.Marshal(global)
-		if err != nil {
-			return err
-		}
-		_, err = file.Write(globalInfo)
-		if err != nil {
-			return err
-		}
-	}
-	return err
 }
 
 func initChannel() {
